requirements.txt
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dash==2.0.0
dash-bootstrap-components==1.0.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.13
Flask==2.0.2
Flask-Compress==1.10.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.0
packaging==21.3
pandas==1.3.5
plotly==5.5.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.0
requests==2.26.0
six==1.16.0
tenacity==8.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
